Rename status as viewer
